+++ adapter_scanner/scanner_utils.py
