+++ library_scanner/adapter.py
