+++ library_scanner/bcd325p2_command_library.py
@@ -1,25 +1,26 @@
-import time
-import scanner_library.bc125atCommandLibrary
-from utilities.scanner_utils import send_command  # Correct location for send_command
-from scanner_adapters.base_adapter import BaseScannerAdapter
-from scanner_library.bc125atCommandLibrary import commands  # Import commands from the correct library
+from utilities.scanner_utils import send_command
+from adapter_scanner.base_adapter import BaseScannerAdapter
+from library_scanner.commands.bcd325p2_commands import commands  # Import commands from the correct library
 
-class BC125ATAdapter(BaseScannerAdapter):
-    def __init__(self, machine_mode=False):
-        self.machine_mode = machine_mode
+import library_scanner.bcd325p2CommandLibrary
+import time
+
+class BC325P2Adapter(BaseScannerAdapter):
+    def __init__(self, machineMode=False):
+        self.machineMode = machineMode
 
     def feedback(self, success, message):
-        if self.machine_mode:
+        if self.machineMode:
             return "OK" if success else "ERROR"
         return message
 
-    def get_help(self, command):
+    def getHelp(self, command):
         try:
-            return scanner_library.bc125atCommandLibrary.getHelp(command)
+            return library_scanner.bc125atCommandLibrary.getHelp(command)
         except Exception as e:
-            return self.feedback(False, f"❌\t[get_help Error] {e}")
+            return self.feedback(False, f"❌\t[getHelp Error] {e}")
 
-    def dump_memory_to_file(self, ser, filename="memorydump.txt", start=0x00000000, end=0x0000FFFF, step=16):
+    def dumpMemoryToFile(self, ser, filename="memorydump.txt", start=0x00000000, end=0x0000FFFF, step=16):
         import sys
 
         def hex32(i): return f"{i:08X}"
@@ -93,7 +94,7 @@ class BC125ATAdapter(BaseScannerAdapter):
         except Exception as e:
             return self.feedback(False, f"❌ [enter_quick_frequency_hold Error] {e}")
 
-    def write_key_beep(self, ser, level=99, lock=0):
+    def writeKeyBeep(self, ser, level=99, lock=0):
         try:
             send_command(ser, "PRG")
             cmd = commands["KBP"]
@@ -103,7 +104,7 @@ class BC125ATAdapter(BaseScannerAdapter):
         except Exception as e:
             return self.feedback(False, f"[writeKeyBeep Error] {e}")
 
-    def read_volume(self, ser):
+    def readVolume(self, ser):
         try:
             response = send_command(ser, commands["VOL"].buildCommand())
             _, value = response.split(",", 1)
@@ -112,7 +113,7 @@ class BC125ATAdapter(BaseScannerAdapter):
         except Exception as e:
             return self.feedback(False, f"❌ [readVolume Error] {e}")
 
-    def write_volume(self, ser, value):
+    def writeVolume(self, ser, value):
         try:
             if not (0.0 <= value <= 1.0):
                 return self.feedback(False, "⚠️ Volume must be between 0.0 and 1.0")
@@ -122,7 +123,7 @@ class BC125ATAdapter(BaseScannerAdapter):
         except Exception as e:
             return self.feedback(False, f"❌ [writeVolume Error] {e}")
 
-    def read_squelch(self, ser):
+    def readSquelch(self, ser):
         try:
             response = send_command(ser, commands["SQL"].buildCommand())
             _, value = response.split(",", 1)
@@ -131,7 +132,7 @@ class BC125ATAdapter(BaseScannerAdapter):
         except Exception as e:
             return self.feedback(False, f"❌ [readSquelch Error] {e}")
 
-    def write_squelch(self, ser, value):
+    def writeSquelch(self, ser, value):
         try:
             if not (0.0 <= value <= 1.0):
                 return self.feedback(False, "❌ Squelch must be between 0.0 and 1.0")
@@ -143,7 +144,7 @@ class BC125ATAdapter(BaseScannerAdapter):
         except Exception as e:
             return self.feedback(False, f"❌ [writeSquelch Error] {e}")
 
-    def read_frequency(self, ser):
+    def readFrequency(self, ser):
         try:
             response = send_command(ser, "PWR")
             parts = response.strip().split(",")
@@ -154,7 +155,7 @@ class BC125ATAdapter(BaseScannerAdapter):
         except Exception as e:
             return self.feedback(False, f"❌ [readFrequency Error] {e}")
 
-    def write_frequency(self, ser, freq_mhz):
+    def writeFrequency(self, ser, freq_mhz):
         try:
             send_command(ser, "PRG")
             send_command(ser, "EPG")
@@ -173,7 +174,7 @@ class BC125ATAdapter(BaseScannerAdapter):
         except Exception as e:
             return self.feedback(False, f"❌ [writeFrequency Error] {e}")
 
-    def read_rssi(self, ser):
+    def readRSSI(self, ser):
         try:
             response = send_command(ser, "PWR")
             parts = response.strip().split(",")
@@ -184,12 +185,12 @@ class BC125ATAdapter(BaseScannerAdapter):
         except Exception as e:
             return self.feedback(False, f"❌ [readRSSI Error] {e}")
 
-    def send_key(self, ser, key_seq):
-        if not key_seq:
+    def sendKey(self, ser, keySeq):
+        if not keySeq:
             return self.feedback(False, "❌ No key(s) provided.")
 
         responses = []
-        for char in key_seq:
+        for char in keySeq:
             if char not in "0123456789<>^.EMFHSLP":
                 responses.append(f"{char} → skipped (invalid key)")
                 continue
@@ -200,19 +201,19 @@ class BC125ATAdapter(BaseScannerAdapter):
                 responses.append(f"❌ {char} → ERROR: {e}")
         return "\n".join(responses)
 
-    def read_model(self, ser):
+    def readModel(self, ser):
         try:
             model = send_command(ser, commands["MDL"].buildCommand())
             return self.feedback(True, f"✅ Model: {model}")
         except Exception as e:
             return self.feedback(False, f"❌ [readModel Error] {e}")
 
-    def read_sw_ver(self, ser):
+    def readSWVer(self, ser):
         try:
             version = send_command(ser, commands["VER"].buildCommand())
             return self.feedback(True, f"✅ Software Version: {version}")
         except Exception as e:
             return self.feedback(False, f"❌ [readSWVer Error] {e}")
 
-    def read_s_meter(self, ser):
+    def readSMeter(self, ser):
         return self.feedback(False, "❌ SMeter not supported")
