+++ main.py
@@ -3,11 +3,11 @@ import logging
 import serial
 import argparse
 from utilities.log_trim import trim_log_file
-from scanner_adapters.scanner_utils import find_all_scanner_ports
-from scanner_adapters.adapter_bc125at import BC125ATAdapter
-from scanner_adapters.adapter_bcd325p2 import BCD325P2Adapter
-from scanner_adapters.adapter_sds100 import SDS100Adapter
-from scanner_adapters.adapter_ardv1 import AORDV1Adapter
+from adapter_scanner.scanner_utils import find_all_scanner_ports
+from adapter_scanner.adapter_bc125at import BC125ATAdapter
+from adapter_scanner.adapter_bcd325p2 import BCD325P2Adapter
+from adapter_scanner.adapter_sds100 import SDS100Adapter
+from adapter_scanner.adapter_ardv1 import AORDV1Adapter
 from command_registry import build_command_table
 from utilities.readlineSetup import initialize_readline
 
@@ -35,7 +35,7 @@ if os.path.getsize("scanner_tool.log") > 10 * 1024 * 1024:  # 10 MB limit
 # ------------------------------------------------------------------------------
 
 # Maps scanner model names to their respective adapter classes
-SCANNER_ADAPTERS = { # if your scanner adapter is not listed here, add it here
+adapter_scanner = { # if your scanner adapter is not listed here, add it here
     "BC125AT": BC125ATAdapter(),
     "BCD325P2": BCD325P2Adapter(),  # Ensure this is correctly instantiated
     "SDS100": SDS100Adapter(),
@@ -197,7 +197,7 @@ def main():
     # Use the scanner model to select the appropriate adapter
     try:
         with serial.Serial(port, 115200, timeout=1) as ser:
-            adapter = SCANNER_ADAPTERS.get(scanner_model)
+            adapter = adapter_scanner.get(scanner_model)
             if not adapter:
                 print(f"No adapter implemented for {scanner_model}.")
                 return
