--- vulture_results.txt
@@ -1,49 +1,0 @@
-scanner_adapters\bc125atAdapter.py:21: unused method 'get_help' (60% confidence)
-scanner_adapters\bc125atAdapter.py:27: unused method 'dump_memory_to_file' (60% confidence)
-scanner_adapters\bc125atAdapter.py:101: unused method 'write_key_beep' (60% confidence)
-scanner_adapters\bc125atAdapter.py:111: unused method 'read_volume' (60% confidence)
-scanner_adapters\bc125atAdapter.py:120: unused method 'write_volume' (60% confidence)
-scanner_adapters\bc125atAdapter.py:130: unused method 'read_squelch' (60% confidence)
-scanner_adapters\bc125atAdapter.py:139: unused method 'write_squelch' (60% confidence)
-scanner_adapters\bc125atAdapter.py:151: unused method 'read_frequency' (60% confidence)
-scanner_adapters\bc125atAdapter.py:162: unused method 'write_frequency' (60% confidence)
-scanner_adapters\bc125atAdapter.py:181: unused method 'read_rssi' (60% confidence)
-scanner_adapters\bc125atAdapter.py:192: unused method 'send_key' (60% confidence)
-scanner_adapters\bc125atAdapter.py:208: unused method 'read_model' (60% confidence)
-scanner_adapters\bc125atAdapter.py:215: unused method 'read_sw_ver' (60% confidence)
-scanner_adapters\bc125atAdapter.py:222: unused method 'read_s_meter' (60% confidence)
-scanner_adapters\bcd325p2Adapter.py:151: unused method 'writeKeyBeep' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:12: unused class 'UnidenAdapter' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:21: unused method 'get_help' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:27: unused method 'dump_memory_to_file' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:282: unused method 'write_key_beep' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:292: unused method 'read_volume' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:301: unused method 'write_volume' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:311: unused method 'read_squelch' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:320: unused method 'write_squelch' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:332: unused method 'read_frequency' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:343: unused method 'write_frequency' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:362: unused method 'read_rssi' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:373: unused method 'send_key' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:389: unused method 'read_model' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:396: unused method 'read_sw_ver' (60% confidence)
-scanner_adapters\scanner_utils_uniden.py:403: unused method 'read_s_meter' (60% confidence)
-scanner_gui\gui\scannerGui.py:152: unused method 'onPortSelected' (60% confidence)
-scanner_gui\gui\scannerGui.py:199: unused method 'knobScrolled' (60% confidence)
-scanner_gui\gui\scannerGui.py:206: unused method 'knobPressed' (60% confidence)
-scanner_gui\scannerUtils.py:34: unused method 'parseResponse' (60% confidence)
-scanner_library\aordv1CommandLibrary.py:49: unused method 'parseResponse' (60% confidence)
-scanner_library\aordv1CommandLibrary.py:55: unused function 'listCommands' (60% confidence)
-scanner_library\bc125atCommandLibrary.py:332: unused function 'listCommands' (60% confidence)
-scanner_library\bcd325p2_command_library.py:98: unused method 'writeKeyBeep' (60% confidence)
-scanner_library\bcd325p2CommandLibrary.py:162: unused function 'listCommands' (60% confidence)
-scanner_library\sds100CommandLibrary.py:1: unused import 'textwrap' (90% confidence)
-scanner_library\sds100CommandLibrary.py:57: unused method 'parseResponse' (60% confidence)
-scanner_library\sds100CommandLibrary.py:63: unused function 'listCommands' (60% confidence)
-scanner_library\uniden_tone_lut.py:25: unused function 'validate_ctcss' (60% confidence)
-scanner_library\uniden_tone_lut.py:48: unused function 'validate_dcs' (60% confidence)
-scannerAdapters\bc125atAdapter.py:99: unused method 'writeKeyBeep' (60% confidence)
-utilities\command_library.py:32: unused method 'parseResponse' (60% confidence)
-utilities\commandLibrary.py:32: unused method 'parseResponse' (60% confidence)
-utilities\scanner_utils.py:254: unused method 'writeKeyBeep' (60% confidence)
-utilities\shared_utils.py:28: unused method 'parseResponse' (60% confidence)
